+++ origins/category/demoCategory.php
@@ -6,6 +6,7 @@ use Exception;
 use srag\Plugins\Hub2\Exception\BuildObjectsFailedException;
 use srag\Plugins\Hub2\Exception\ConnectionFailedException;
 use srag\Plugins\Hub2\Exception\ParseDataFailedException;
+use srag\Plugins\Hub2\Log\ILog;
 use srag\Plugins\Hub2\Object\Category\CategoryDTO;
 use srag\Plugins\Hub2\Object\DTO\IDataTransferObject;
 use srag\Plugins\Hub2\Object\HookObject;
@@ -115,9 +116,9 @@ class demoCategory extends AbstractOriginImplementation {
 	 *
 	 * Note that if you do not throw any of the exceptions above, the sync will continue.
 	 *
-	 * @param Exception $e
+	 * @param ILog $log
 	 */
-	public function handleException(Exception $e) { }
+	public function handleLog(ILog $log) { }
 
 
 	/**