+++ src/Sync/Summary/OriginSyncSummaryBase.php
@@ -106,7 +106,7 @@ abstract class OriginSyncSummaryBase implements IOriginSyncSummary {
 		$msg = "";
 		if (!$only_logs) {
 			// Print out some useful statistics: --> Should maybe be a OriginSyncSummary object
-			$msg .= self::plugin()->translate("summary_for", "", [ $originSync->getOrigin()->getTitle() ]) . "\n**********\n";
+			$msg .= self::plugin()->translate("summary_for", "", [ $originSync->getOrigin()->getTitle() ]) . "\n";
 			$msg .= self::plugin()->translate("summary_delivered_data_sets", "", [ $originSync->getCountDelivered() ]) . "\n";
 			$msg .= self::plugin()->translate("summary_failed", "", [ $originSync->getCountProcessedByStatus(IObject::STATUS_FAILED) ]) . "\n";
 			$msg .= self::plugin()->translate("summary_created", "", [ $originSync->getCountProcessedByStatus(IObject::STATUS_CREATED) ]) . "\n";